(refac): split the datasets into their own modules
