+++ src/careamics/dataset/__init__.py
@@ -2,20 +2,16 @@
 
 __all__ = [
     "InMemoryDataset",
-    "InMemoryPredictionDataset",
+    "InMemoryPredDataset",
     "InMemoryTiledPredictionDataset",
     "PathIterableDataset",
     "IterableTiledPredictionDataset",
     "IterablePredictionDataset",
 ]
 
-from .in_memory_dataset import (
-    InMemoryDataset,
-    InMemoryPredictionDataset,
-    InMemoryTiledPredictionDataset,
-)
-from .iterable_dataset import (
-    IterablePredictionDataset,
-    IterableTiledPredictionDataset,
-    PathIterableDataset,
-)
+from .in_memory_dataset import InMemoryDataset
+from .in_memory_pred_dataset import InMemoryPredDataset
+from .in_memory_tiled_pred_dataset import InMemoryTiledPredictionDataset
+from .iterable_dataset import PathIterableDataset
+from .iterable_pred_dataset import IterablePredictionDataset
+from .iterable_tiled_pred_dataset import IterableTiledPredictionDataset
