+++ src/careamics/dataset/dataset_utils/__init__.py
@@ -8,11 +8,13 @@ __all__ = [
     "read_tiff",
     "get_read_func",
     "read_zarr",
+    "iterate_over_files",
 ]
 
 
 from .dataset_utils import reshape_array
 from .file_utils import get_files_size, list_files, validate_source_target_files
+from .iterate_over_files import iterate_over_files
 from .read_tiff import read_tiff
 from .read_utils import get_read_func
 from .read_zarr import read_zarr
